Refactored TagListView
